Postfix

+++ src/ast/tuple_index_node.h
@@ -8,10 +8,10 @@ namespace og {
 
   class tuple_index_node: public cdk::lvalue_node {
     cdk::expression_node *_base;
-    int _index;
+    size_t _index;
 
   public:
-    tuple_index_node(int lineno, cdk::expression_node *base, int index) :
+    tuple_index_node(int lineno, cdk::expression_node *base, size_t index) :
         cdk::lvalue_node(lineno), _base(base), _index(index) {
     }
 
@@ -19,7 +19,7 @@ namespace og {
     cdk::expression_node *base() {
       return _base;
     }
-    int index() {
+    size_t index() {
       return _index;
     }
